docker compose and exceptions

+++ src/main/java/com/stackroute/fullstackbackend/service/UserServiceImpl.java
@@ -24,12 +24,12 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User addUser(User user) {
+    public User addUser(User user) throws UserAlreadyExistsException {
         User addedUser=null;
-        User username=userRepo.existsByName(user.getUsername());
-//        if(username.getName()==user.getName()){
-//            throw new UserAlreadyExistsException("User Already Exists");
-//        }
+        User existingUser=userRepo.existsByName(user.getUsername());
+        if(existingUser.getName()==user.getName()){
+            throw new UserAlreadyExistsException("User Already Exists");
+        }
             addedUser=userRepo.save(user);
 
 
@@ -37,23 +37,32 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<User> getAllUsers() {
+    public List<User> getAllUsers() throws UserNotFoundException{
+
         List<User> users= (List<User>) userRepo.getAllUsers();
+        if(users== null||users.isEmpty()){
+            throw new UserNotFoundException("No users are present");
+        }
         return users;
     }
 
 
     @Override
 
-    public boolean addFriendByName(String username1,String username2){
+    public boolean addFriendByName(String username1,String username2) throws UserNotFoundException{
 
 
-            System.out.println("the parameter data is  "+username1+"  "+username2);
+//            System.out.println("the parameter data is  "+username1+"  "+username2);
+
+
 
             User friend = userRepo.makeFriend(username1, username2);
 
            User u1= userRepo.existsByName(username1);
            User u2= userRepo.existsByName(username2);
+           if(u2==null){
+               throw new UserNotFoundException("sorry..your friend doesnt exist");
+           }
            u1.setFriends(u2.getId());
            u2.setFriends(u1.getId());
            userRepo.save(u1);
@@ -83,11 +92,11 @@ public class UserServiceImpl implements UserService {
 //
 //    }
 
-    public boolean deleteUserByUsername(String username)  {
+    public boolean deleteUserByUsername(String username) throws UserNotFoundException{
         User deleteuser=userRepo.existsByName(username);
-//        if(deleteuser.getUsername()!=username){
-//            throw new UserNotFoundException("User Not Exists");
-//        }
+        if(deleteuser.getUsername()!=username){
+            throw new UserNotFoundException("User Not Exists");
+        }
         deleteuser=userRepo.deleteUserByUsername(username);
         return true;
     }
@@ -109,7 +118,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<User> getUserFriends(String username) {
+    public List<User> getUserFriends(String username)  {
    List<Long> lists= userRepo.existsByName(username).getFriends();
         List<User> friendslist=new ArrayList<>() ;
 
@@ -121,27 +130,35 @@ public class UserServiceImpl implements UserService {
         return friendslist;
     }
 @Override
-    public User getUserDetails(String username) {
+    public User getUserDetails(String username) throws UserNotFoundException{
         User userdetails= userRepo.getUserDetails(username);
-
+        if(userdetails==null){
+            throw new UserNotFoundException("your details are not updated");
+        }
         return userdetails;
 
     }
     @Override
-    public List<User> searchUsersByName(String input){
+    public List<User> searchUsersByName(String input) throws UserNotFoundException{
         List<User> userList=(List)userRepo.getAllUsers();
         List<User> matchedList=new ArrayList<>();
         for(int i=0;i<userList.size();i++){
             if(userList.get(i).getName().toLowerCase().matches("(.*)"+input.toLowerCase()+"(.*)"))
                 matchedList.add(userList.get(i));
         }
+        if(matchedList.isEmpty()){
+            throw new UserNotFoundException("No users with this name found");
+        }
         return matchedList;
     }
 
     @Override
-    public User deleteUserFriendsByName(String username1, String username2) {
+    public User deleteUserFriendsByName(String username1, String username2) throws UserNotFoundException{
         User loggedInUser = userRepo.deleteUserfriendsByName(username1, username2);
         User friend=userRepo.existsByName(username2);
+        if(friend==null){
+            throw new UserNotFoundException("You dont have a friend with this name");
+        }
         Long id=loggedInUser.getId();
         friend.friends.remove(id);
         loggedInUser.friends.remove(friend.getId());
